+++ powercli/__init__.py
@@ -1,7 +1,6 @@
 """PowerCLI - Build powerful command-line applications in Python."""
 
 # TODO: __main__ to generate completions
-# TODO: example usage for (sub)commands (see also: Bun CLI)
 
 from __future__ import annotations